AoC.Tests/Day07/Day7SolverTests.cs
using AoC.Day07;

namespace AoC.Tests.Day07;

public class Day7SolverTests
{
    private readonly Day7Solver _sut = new();

    private const string ExampleInput = """
        32T3K 765
        T55J5 684
        KK677 28
        KTJJT 220
        QQQJA 483
        """;

    [Test]
    public void Part1Example()
    {
        // ACT
        var part1ExampleResult = _sut.SolvePart1(ExampleInput);

        // ASSERT
        part1ExampleResult.Should().Be(6440);
    }

    [Test]
    public void Part1ReTest()
    {
        // ACT
        var part1Result = _sut.SolvePart1();

        // ASSERT
        part1Result.Should().Be(248217452);
    }

    [Test]
    public void Part2Example()
    {
        // ACT
        var part2ExampleResult = _sut.SolvePart2(ExampleInput);

        // ASSERT
        part2ExampleResult.Should().Be(5905);
    }

    [Test]
    public void Part2ReTest()
    {
        // ACT
        var part2Result = _sut.SolvePart2();

        // ASSERT
        part2Result.Should().BeGreaterThan(245147007);
        part2Result.Should().BeGreaterThan(245406987);        
        part2Result.Should().Be(245576185);
    }
}
